Improves metadata parsing for PR 31763 (#32067)
* Description: uses regex instead of tmp file.

* Adding test to check for like tag before and after target.

* Description: changes regex strategy.

Why?

Lookahead was not constrained with each element and thus picked up
<meta from one and then if not a match, grabbed the name and content
from another upstream.

The new strategy parses all meta elements with a content
attribute. Then loops through them to find the description element.

Why this order?
The content attribute can contain HTML tags. The > or /> symbol is
matched as the end of the meta element (it's closing symbol). If
this happens, the content is truncated. Boo.

Switching the parsing order solves this problem.

Bonus: allows for pre-parsing of all meta elements. Performance boost.

* Refactors getting meta with content elements for reuse.

* Improves getting <head>..</head> element.

- Isolates to the only the <head>..</head> element by stripping
  all content before the opening tag and ensuring it includes a closing
  </head> tag.
- Performance improvements:
   - Bails out early if no opening tag is found.
   - Uses native string functions instead of regex.

* Image: use same parsing strategy as description.

* Refactor to reuse the process for getting the metadata from the list of meta elements.

* Convert description HTML entities into HTML.